# app/docs/swagger.yaml
definitions:
...
  payementmethod.PayementMethod:
    properties:
      name:
        example: Credit Card
        type: string
      payement_method_id:
        example: 1
        type: integer
    type: object
...
  sale.CreateSaleInput:
    properties:
      lines:
        items:
          $ref: '#/definitions/sale.SaleLine'
        minItems: 1
        type: array
      payement_method_id:
        example: 1
        type: integer
    required:
    - lines
    - payement_method_id
    type: object
  sale.Sale:
    properties:
      is_deleted:
        type: boolean
      lines:
        items:
          $ref: '#/definitions/sale.SaleLine'
        type: array
      payement_method:
        $ref: '#/definitions/payementmethod.PayementMethod'
      payement_method_id:
        type: integer
      price:
        type: number
      sale_id:
        type: integer
      time_stamp:
        type: string
    type: object
  sale.SaleLine:
    properties:
      item_id:
        example: 1
        type: integer
      name:
        example: Coffee
        type: string
      quantity:
        example: 2
        minimum: 1
        type: integer
      tax_rate:
        example: 0.2
        type: number
      unit_price:
        example: 4.5
        type: number
    required:
    - item_id
    - quantity
    - unit_price
    type: object
  sale.UpdateSaleInput:
    properties:
      lines:
        items:
          $ref: '#/definitions/sale.UpdateSaleLine'
        minItems: 1
        type: array
      payement_method_id:
        example: 2
        type: integer
    type: object
  sale.UpdateSaleLine:
    properties:
      item_id:
        example: 2
        type: integer
      name:
        example: Tea
        type: string
      quantity:
        example: 1
        minimum: 0
        type: integer
    required:
    - item_id
    - quantity
    type: object
  salehistory.SaleHistory:
    properties:
      changed_at:
        type: string
      changed_by_profile_id:
        type: integer
      changes:
        additionalProperties: {}
        type: object
      history_id:
        type: integer
      is_deleted:
        type: boolean
      lines:
        items:
          $ref: '#/definitions/salehistory.SaleLineSnapshot'
        type: array
      payement_method_id:
        type: integer
      price:
        type: number
      sale_id:
        type: integer
      time_stamp:
        type: string
    type: object
  salehistory.SaleLineSnapshot:
    properties:
      item_id:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      tax_rate:
        type: number
      unit_price:
        type: number
    type: object
...
info:
  contact: {}
  title: Tili API
  version: "0.1"
paths:
...
  /sales:
    get:
      description: Returns all non-deleted sales ordered by most recent first.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sale.Sale'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all sales
      tags:
      - sales
    post:
      consumes:
      - application/json
      description: Creates a sale with one or more line items. The total is computed
        from unit prices × quantities. Returns 400 if the computed total is not positive.
      parameters:
      - description: Sale payload
        in: body
        name: sale
        required: true
        schema:
          $ref: '#/definitions/sale.CreateSaleInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sale.Sale'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new sale
      tags:
      - sales
  /sales/{id}:
    delete:
      description: Soft-deletes a sale by setting is_deleted=true. The deletion is
        recorded in sale_history. Requires manager access.
      parameters:
      - description: Sale ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a sale
      tags:
      - sales
    get:
      description: Returns a single non-deleted sale by its ID.
      parameters:
      - description: Sale ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sale.Sale'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a sale by ID
      tags:
      - sales
    put:
      consumes:
      - application/json
      description: 'Partially updates a sale. Only the fields provided are changed.
        For lines: existing item IDs update their quantity; new item IDs are added
        (name and unit_price required); items not mentioned are kept unchanged. The
        total is recomputed. The change is recorded in sale_history. Requires manager
        access.'
      parameters:
      - description: Sale ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: sale
        required: true
        schema:
          $ref: '#/definitions/sale.UpdateSaleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sale.Sale'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a sale
      tags:
      - sales
  /sales/{id}/history:
    get:
      description: Returns the audit trail of changes for a sale, ordered most-recent-first.
        Each row captures the state of the sale after the change was applied, alongside
        a diff of what changed (fields "action", and any modified field with from/to
        values). Requires manager access.
      parameters:
      - description: Sale ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/salehistory.SaleHistory'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List the change history of a sale
      tags:
      - sales
...
securityDefinitions:
  AccountToken:
    description: JWT obtenu après login account (POST /account/login)
    in: header
    name: Authorization
    type: apiKey
  ProfileToken:
    description: JWT obtenu après login profil avec PIN (POST /profile/login/pin)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"